refactor: split tauri command modules
