+++ desktop/src-tauri/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod diagnostics;
+pub(crate) mod profiles;
+pub(crate) mod runtime;
